backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
httpx>=0.27.0
python-dotenv>=1.0.1
langgraph>=0.2.0
google-genai>=0.7.0
pillow>=10.4.0
rembg>=2.0.60
pytest>=8.3.0

